LayoutTests/imported/w3c/web-platform-tests/webmessaging/postMessage_MessagePorts_sorigin-expected.txt


PASS Test Description: postMessage to same-origin iframe with MessagePort array containing 100 ports.
PASS MessageChannel is supported.
